w5-Assignment/requirements.txt
# Core LLM & Agents
langchain
langchain-community
langchain_huggingface
google-generativeai
torch
torchvision
torchaudio
huggingface-hub
transformers

# Vector Storage For Journal Entries
faiss-cpu

# Embedding and NLP Tools
sentence-transformers
nltk
textblob

# Frontend and Backend
streamlit
flask

# Data Handling and Visualization
pandas
matplotlib

# Utilities
python-dotenv
tqdm
rich
config